+++ sphinx/source/changelog.rst
@@ -7,6 +7,7 @@ Changelog
   into the memory snapshot)
 * Added the :ref:`defw` directive (for inserting word values into the memory
   snapshot)
+* Added the :ref:`remote` directive (for creating a remote entry)
 * Added support to the :ref:`Resources` section for specifying files using
   wildcard characters (``*``, ``?`` and ``[]``)
 * Added support for parsing block-level comments that are not left-padded by a